day5: part1

+++ src/challenges/day02/mod.rs
@@ -2,7 +2,7 @@ use std::time::Instant;
 
 const INPUT: &str = include_str!("input2.txt");
 
-pub fn run() -> ((u32, f64), (u32, f64)) {
+pub fn run() -> ((u128, f64), (u128, f64)) {
   let start_part1 = Instant::now();
   let result_part1 = part1();
   let duration_part1 = start_part1.elapsed();
@@ -20,20 +20,21 @@ pub fn run() -> ((u32, f64), (u32, f64)) {
 }
 
 // red, green, blue
-const LIMITS: [(usize, u32); 3] = [(0, 12), (1, 13), (2, 14)];
+const LIMITS: [(usize, u128); 3] = [(0, 12), (1, 13), (2, 14)];
 
-fn part1() -> u32 {
+fn part1() -> u128 {
   INPUT
     .lines()
     .filter_map(|line| {
       let (game_id_str, colors) = line.split_once(": ")?;
-      let game_id: u32 = game_id_str.split_whitespace().last()?.parse().ok()?;
+      let game_id: u128 =
+        game_id_str.split_whitespace().last()?.parse().ok()?;
       for set in colors.split(';') {
         let mut counts = [0; 3];
 
         for color in set.split(',').map(str::trim) {
           let mut parts = color.split_whitespace();
-          let count: u32 = parts.next().unwrap().parse().unwrap_or(0);
+          let count: u128 = parts.next().unwrap().parse().unwrap_or(0);
           let color_index = match parts.next().unwrap() {
             "red" => 0,
             "green" => 1,
@@ -53,7 +54,7 @@ fn part1() -> u32 {
     .sum()
 }
 
-fn part2() -> u32 {
+fn part2() -> u128 {
   INPUT
     .lines()
     .map(|line| {
@@ -66,7 +67,7 @@ fn part2() -> u32 {
 
         for color in set.split(',').map(str::trim) {
           let mut parts = color.split_whitespace();
-          let count: u32 = parts.next().unwrap().parse().unwrap_or(0);
+          let count: u128 = parts.next().unwrap().parse().unwrap_or(0);
           let color_index = match parts.next().unwrap() {
             "red" => 0,
             "green" => 1,
